Allow // comments anywhere in builtins.txt; simplify Makefile
Removes lines that start with // at the very start of the line during sed processing, to allow for comments, as opposed to blindly deleting the first line.

Also merge the sed calls into one, use single instead of double quotes, and remove builtins_txt.cc in case sed fails, so that a subsequent 'make' invocation doesn't work on a bogus or incomplete builtins_txt.cc file.