\t stands for 4 spaces.
This commit is somewhat wasteful in that it reserves space as if the whole string was made of \t, but that's the worst case and has to be contemplated. The alternative is to make two passes over the string, one to calculate the final size and another to extract the characters. We've chosen the easy way.