Allow state changes where they are really allowed, and fix messages.
They have different effects for Mono and LSO, because in Mono the stack can't be corrupted. Also, state changes are DISallowed when the 'if' has an 'else', but are allowed in a 'while', 'do' or 'for' too.

Separate the test case for this into LSO and Mono. There was no test in the suite for E20005 anyway.