Support condition always false in `for` statements.
Adds an expression node to the init and cont expressions of the `for` loop, so that both have their corresponding expressions as children rather than siblings.

Since the value of multiple simultaneous child expressions with no operator is undefined in LSL (one could argue it's the value of the last node, as in C, but it has no use anyway), it's only propagated if there's only one child.