Fix line counter on strings that have newlines.
Multiline strings were erroneously counted as being single-line, for error reporting purposes. This was causing the error positions after the string to be shifted as many lines as newlines had the multiline string(s).

The test suite does not have a provision to test this, therefore no test is added.