Report correct column in UTF-8 source programs.
LSL ignores a bunch of characters in the source: $, `, ', \, ?, #, control characters including DEL, and UTF-8 characters with a Unicode codepoint >= U+0080.

When such a character is present in a line outside a string, it is ignored as if it was a space. But the line counter counts bytes, therefore it doesn't count right when the characters are UTF-8 multibyte characters.

Example:

key »k« = ;

reports:

ERROR:: (  1, 13): syntax error, unexpected ';'

and the column is out of range. With this patch it reports:

ERROR:: (  1, 11): syntax error, unexpected ';'

which is the column where the ';' is.