Assign type of node based on operator in some cases
Some operators have a fixed return type. In these cases, to prevent a cascade of invalid operation errors, assign the return type of the operator, rather than that of the first operand.

That type happens to be integer in all cases handled.

This is especially important in the case of >= and <= because that's parsed as !(a < b) and !(a > b) respectively, and if there's a type mismatch in the inner operation, another error is raised in the ! operation, potentially confusing the user.